+++ features/admin.private-key-jwt.v1/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @wso2is/admin.private-key-jwt.v1
 
+## 2.20.48
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.core.v1@2.21.18
+  - @wso2is/admin.server-configurations.v1@2.20.48
+
 ## 2.20.47
 
 ### Patch Changes