+++ features/common.ai.v1/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @wso2is/common.ai.v1
 
+## 2.20.48
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.core.v1@2.21.18
+
 ## 2.20.47
 
 ### Patch Changes